refact(Error): rename Error to BusinessError
