+++ src/domains/shared/ValidationResponse.ts
@@ -1,9 +1,9 @@
-import Error from "@/domains/shared/exceptions/Error";
+import BusinessError from "@/domains/shared/exceptions/BusinessError";
 
 class ValidationResponse {
-  constructor(public ok: boolean, public errors: Error[] = []) {}
+  constructor(public ok: boolean, public errors: BusinessError[] = []) {}
 
-  addError(error: Error) {
+  addError(error: BusinessError) {
     this.errors.push(error);
     this.ok = false;
   }
